+++ tests/test_simple_heat_source.py
@@ -1,7 +1,7 @@
 """Test for generic heat source."""
 import pytest
 from hisim import component as cp
-from hisim.components import simple_const_heat_source
+from hisim.components import simple_heat_source
 from hisim.simulationparameters import SimulationParameters
 
 
@@ -14,7 +14,7 @@ def test_heat_source():
     my_simulation_parameters = SimulationParameters.one_day_only(2017, seconds_per_timestep)
 
     # default config
-    my_heat_source_config = simple_const_heat_source.SimpleHeatSourceConfig.get_default_config_const_power()
+    my_heat_source_config = simple_heat_source.SimpleHeatSourceConfig.get_default_config_const_power()
 
     # definition of outputs
     number_of_outputs = 1
@@ -22,7 +22,7 @@ def test_heat_source():
 
     # ===================================================================================================================
     # Set Heat Pump
-    my_heat_source = simple_const_heat_source.SimpleHeatSource(
+    my_heat_source = simple_heat_source.SimpleHeatSource(
         config=my_heat_source_config, my_simulation_parameters=my_simulation_parameters
     )
